pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.38.42",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.66",
    "langchain-openai>=0.3.27",
    "langchain-pinecone>=0.2.8",
    "langchain-voyageai>=0.1.6",
    "langgraph>=0.5.0",
    "pandas>=2.3.0",
    "pinecone>=7.2.0",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=20.0.0",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.46.1",
    "tiktoken>=0.9.0",
    "voyageai>=0.3.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "python-dotenv>=1.1.1",
]
